Fix GPU PWR k_inf: rank mismatch + URR total recomputation
Two bugs were making GPU k_inf 4900 pcm below CPU reference for PWR pin cell:

1. Discrete-level SVD rank mismatch (xs_provider.rs)
   Discrete levels were built with svd_rank.min(2)=2, but the GPU kernel
   uses the top-level P_RANK=5 as basis stride when calling
   svd_reconstruct(basis, coeffs, e_idx, rank). This caused reads past
   the end of each level's basis into the next level's data, making
   per-level XS essentially random. Now uses the full svd_rank so stride
   matches.

2. URR total XS not recomputed (transport.cu)
   After apply_urr() modifies s_el, s_fis, s_cap, micro_t was left
   unchanged, leaving the macroscopic total inconsistent with sum of
   partials. Collision distance and reaction sampling both saw a wrong
   total. Now tracks per-channel delta and adjusts micro_t, matching
   the CPU apply_urr which recomputes xs.total.

Also:
- SVD basis f32 → f64 everywhere (kernel.rs, gpu_transport.rs, gpu.rs,
  transport.cu) to eliminate the known f32 overshoot.
- GPU PWR nuclide specs updated to 9 entries with separate O16 at 900K
  (fuel, idx 2) and 600K (water, idx 8), matching the CPU pwr_pincell
  layout exactly.
- Added CPU-matching hydrogen special case in GPU mu_lab computation
  for A<=1+eps (sqrt((1+mu_cm)/2) instead of the general formula).
- gpu_cpu_trace now uses TableXsProvider on the CPU side to compare
  GPU pointwise against the same pointwise reference.

Result (PWR pin cell, 3 seeds, 10k particles, 40 active batches):
  CPU Table:  1.322 +/- 0.002
  GPU:        1.328 +/- 0.001   (was 1.273, now matches OpenMC reference 1.328)
  Gap:        +60 pcm (within statistics)

All 36 CPU tests still pass.

+++ rust_prototype/src/gpu.rs
@@ -18,7 +18,7 @@ pub mod cuda {
     /// CUDA kernel source for SVD reconstruction.
     const SVD_KERNEL_SRC: &str = r#"
 extern "C" __global__ void svd_reconstruct(
-    const float* __restrict__ basis,
+    const double* __restrict__ basis,
     const double* __restrict__ coeffs,
     double* __restrict__ output,
     const int* __restrict__ energy_indices,
@@ -36,10 +36,10 @@ extern "C" __global__ void svd_reconstruct(
     int e_idx = energy_indices[tid];
     if (e_idx < 0 || e_idx >= n_e) return;
 
-    const float* row = &basis[e_idx * rank];
+    const double* row = &basis[e_idx * rank];
     double acc = 0.0;
     for (int j = 0; j < rank; j++) {
-        acc = fma((double)row[j], shared_coeffs[j], acc);
+        acc = fma(row[j], shared_coeffs[j], acc);
     }
     output[tid] = exp2(acc * 3.321928094887362);
 }
@@ -71,11 +71,11 @@ extern "C" __global__ void svd_reconstruct(
 
         /// Batch SVD reconstruction on GPU.
         ///
-        /// Given a basis matrix (f32), coefficients (f64), and energy indices
+        /// Given a basis matrix (f64), coefficients (f64), and energy indices
         /// for N particles, returns N cross-section values (f64, linear scale).
         pub fn reconstruct_batch(
             &self,
-            basis: &[f32],
+            basis: &[f64],
             coeffs: &[f64],
             energy_indices: &[i32],
             n_e: usize,
